Fix some Solaris related issues, among them a SIGSEGV in t.bulkmode.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@2461 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/lexer.h
@@ -46,7 +46,7 @@ typedef token_t yylex_t(void);
 typedef struct lexer_s {
     yylex_t  *yylex;
     char    **yytext;
-    size_t   *yyleng;
+    int      *yyleng;
 } lexer_t;
 
 extern lexer_t *lexer;
